Handle exception in forward
The very first forward call in the loop
will throw an HttpLinkServerException when
the server resonse is "unauthenticated" and
we don't get a chance to refresh our token.

Handle those exceptions and create an error response
so that the loop body will be able to redo the call
with updated auth headers.